reorg: Extract AbstractUserAttributeMapper from AbstractCASProtocolMapper
