+++ src/main/java/org/keycloak/protocol/cas/mappers/FullNameMapper.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-public class FullNameMapper extends AbstractCASProtocolMapper {
+public class FullNameMapper extends AbstractUserAttributeMapper {
     private static final List<ProviderConfigProperty> configProperties = new ArrayList<ProviderConfigProperty>();
 
     static {
