+++ src/main/java/org/keycloak/protocol/cas/mappers/GroupMembershipMapper.java
@@ -10,7 +10,7 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
-public class GroupMembershipMapper extends AbstractCASProtocolMapper {
+public class GroupMembershipMapper extends AbstractUserAttributeMapper {
     private static final List<ProviderConfigProperty> configProperties = new ArrayList<ProviderConfigProperty>();
 
     private static final String FULL_PATH = "full.path";
